+++ src/libs/constants/inspirationpage.ts
@@ -1,0 +1,3 @@
+export const PRODUCT_IMAGE = ['wallPanel', 'aluminumBox', 'manufacturing', 'paintAndMarking'];
+
+export const a = 'A';
